Computer will hit adjacent squares stupidly
The computer will hit ALL the adjacent squares
even though ship has been sunk or the
orientation has been found.

also fixed a bug where my patrol boat couldnt be
placed on the maximum right side of the board.

//tedious thing i am skipping for now.
I think if want to cut down the squares to hit
if 2 squares are hit, it is possible.

if length of computerMove !== 0 (meaning we already
found the second hit) and opponent[index].hasHIt ===
true

find out the ship we hit two times in a row

then we can find the difference btwn the last 2 index
if it is +1 or -1
then we can clear out the current computer moves.
then we calculate how many boxes to hit until sunk
example: carrier 5 box - 2 (since we hit 2)
so we will have +3 and -3 move and add this 6 move
to the computerMove

then we need to find if length of computer move !== 0
and the opponent[index].hasHit === true but opponent[index]
.hasSHip === false
then we want to continue to hit the according to the moves
we created for it from the comparison of 2 last move

but what if the move hashit and there is a ship.
then we want to continue shooting the moves we put in the
move slots.

we need to find out if the ship we are hitting has sunk
if sunk, then we leave this function and back
to random